requirements.txt

Django==4.0.1
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
djongo==1.3.6
dnspython==2.1.0
drf-yasg==1.20.0
idna==3.3
importlib-metadata==4.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.0.1
packaging==21.3
Pillow==9.0.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
PyJWT==2.3.0
pymongo==3.12.3
pyparsing==3.0.6
pytz==2021.3
requests==2.27.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
simplejwt==2.0.1
sqlparse==0.2.4
typing==3.7.4.3
uritemplate==4.1.1
urllib3==1.26.7
zipp==3.7.0